+++ vortex-io/src/file/source.rs
@@ -3,9 +3,9 @@
 
 use std::sync::Arc;
 
+use futures::FutureExt;
 use futures::future::{BoxFuture, LocalBoxFuture};
 use futures::stream::BoxStream;
-use futures::FutureExt;
 use vortex_error::VortexResult;
 
 use crate::file::request::IoRequest;
